refactor(config): split GameConfig.ts into per-component files (#70)
Split the 578-line src/config/GameConfig.ts into per-component files
(map, player, ai, camera, input, animation, rendering, scene,
server-sync) and turn GameConfig.ts into a pure re-export barrel.
Existing import paths are preserved.
