+++ src/config/input.ts
@@ -1,0 +1,9 @@
+/**
+ * Keyboard key constants for game controls
+ */
+export const KEYBOARD_KEYS = {
+  DANCE: 'f',
+  DANCE_UPPER: 'F',
+  SPECTATOR_TOGGLE: 't',
+  SPECTATOR_TOGGLE_UPPER: 'T',
+} as const;
